requirements.txt
asgiref==3.5.2
cassandra-driver==3.25.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.0
DateTime==4.4
Django==4.0.6
django-admin==2.0.1
django-cors-headers==3.13.0
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-six==1.0.4
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
mysqlclient==2.1.0
Pillow==9.1.0
ply==3.11
psutil==5.9.0
PyJWT==2.4.0
pytz==2022.1
requests==2.28.1
sqlparse==0.4.2
typing-extensions==4.2.0
virtualenv==20.11.0
